refactor(IdentityServer): 重构目录文件
