+++ modules/IdentityServer/src/J3space.Abp.IdentityServer.Application/J3space/Abp/IdentityServer/AbpIdentityServerApplicationModule.cs
@@ -6,8 +6,8 @@ using Volo.Abp.PermissionManagement;
 namespace J3space.Abp.IdentityServer
 {
     [DependsOn(
-        typeof(AbpIdentityServerApplicationContractsModule),
         typeof(AbpAutoMapperModule),
+        typeof(AbpIdentityServerApplicationContractsModule),
         typeof(AbpPermissionManagementApplicationModule)
     )]
     public class AbpIdentityServerApplicationModule : AbpModule
