+++ modules/IdentityServer/src/J3space.Abp.IdentityServer.Application/J3space/Abp/IdentityServer/ClientAppService.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Threading.Tasks;
 using J3space.Abp.IdentityServer.Clients;
 using J3space.Abp.IdentityServer.Clients.Dto;
+using J3space.Abp.IdentityServer.Permissions;
 using Microsoft.AspNetCore.Authorization;
 using Volo.Abp;
 using Volo.Abp.Application.Dtos;
